+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - TT-1917 - Add print summary feature
 - TT-1965 - Content change
 - TT-2001 - Allow only one commodity to be selected
+- TT-2001 - Persist business and personal details
 
 ### Bug fixes